os/selftest/requirements.txt
pytest==8.4.2
testinfra==6.0.0
PyYAML==6.0.2
